requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
anyconfig==0.13.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
comm==0.2.0
datasets==2.14.7
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
docker-pycreds==0.4.0
evaluate==0.4.1
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.2
huggingface-hub==0.17.3
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
lazy_loader==0.3
Levenshtein==0.23.0
lightning==2.1.1
lightning-utilities==0.9.0
lion-pytorch==0.1.2
lit==17.0.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
natsort==8.4.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
nougat-ocr==0.1.17
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pypdf==3.17.1
pypdfium2==4.24.0
python-dateutil==2.8.2
python-Levenshtein==0.23.0
pytorch-lightning==2.1.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sconf==0.2.5
sentencepiece==0.1.99
sentry-sdk==1.37.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.5.4
tokenizers==0.14.1
torch==2.0.0
torchmetrics==1.2.0
torchvision==0.15.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.0
wcwidth==0.2.10
Werkzeug==3.0.1
wheel==0.41.3
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
